+++ authidentity/src/main/java/aivle/domain/enums/UserRole.java
@@ -1,4 +1,4 @@
-package aivle.domain;
+package aivle.domain.enums;
 
 public enum UserRole {
     ADMIN, AUTHOR, USER
