+++ authidentity/src/main/java/aivle/domain/event/SignedUp.java
@@ -1,7 +1,9 @@
-package aivle.domain;
+package aivle.domain.event;
 
 import java.util.Date;
 
+import aivle.domain.entity.AdminAccount;
+import aivle.domain.entity.UserAccount;
 import aivle.infra.AbstractEvent;
 import lombok.Data;
 import lombok.ToString;
